Refactor: Extract OpenAPI setup into a separate module
Moved the OpenAPI setup logic to a dedicated module for better organization. Created 'setup-openapi.ts' for setting up Swagger documentation and 'generate.ts' for handling document generation. Updated main.ts to use the new setup module.
